everything done except for the NIST test the chaotic stream generator is not working properly

+++ src/chaoticStreamProcessing/lorenzStreamProcessor.h
@@ -1,17 +1,17 @@
-
 #pragma once
 #include <cstdint>
 
 class lorenzStreamProcessor {
-    private:
-        int m_size;                  // Target pixel count in bytes
-        uint32_t* m_intStream; // Contiguous 8-bit stream for VRAM
+private:
+    int       m_size;      // Target pixel/byte count
+    uint32_t* m_intStream; // Contiguous 32-bit diffusion stream
 
-    public:
-        lorenzStreamProcessor(int pixelCount);
-        ~lorenzStreamProcessor();
+public:
+    lorenzStreamProcessor(int pixelCount);
+    ~lorenzStreamProcessor();
 
-        uint32_t extractIntegralValues(float floatingValue);
-        void ingestRawStream(const float* rawLorenzStream);
-        uint32_t* getDiffusionValues() const { return m_intStream; }
+    uint32_t extractIntegralValues(double floatingValue);
+    void     ingestRawStream(const double* rawLorenzStream);
+
+    uint32_t* getDiffusionValues() const { return m_intStream; }
 };
